backend/requirements.txt
celery==5.3.6
channels==4.0.0
channels_redis==4.2.0
confluent-kafka==2.3.0
cssutils==2.9.0
daphne==4.1.0
Django==4.2.10 # Do not upgrade until django celery beat supports django 5.0
django-auditlog==2.3.0
django-cacheops==7.0.2
django-celery-beat==2.1.0 # Do not upgrade until they support django 5.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-compressor==4.4
django-cors-headers==4.3.1
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-guardian==2.4.0
django-localflavor==4.0
django-model-utils==4.4.0
django-notifications-hq==1.8.3
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-silk==5.1.0
django-stubs-ext==4.2.7
djangorestframework~=3.14.0
djangorestframework-camel-case==1.4.2
drf-spectacular==0.27.1
drf-standardized-errors[openapi]==0.12.6
geopy==2.4.1
googlemaps==4.10.0
graphene-django==3.2.0
hiredis==2.3.2
openpyxl==3.1.2
pandas==2.2.1
phonenumbers==8.13.31
Pillow==10.2.0
psutil==5.9.8
psycopg[binary]==3.1.18
pyftpdlib==1.5.9
pypdf==4.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
requests==2.31.0
rich==13.7.1
sympy==1.12
watchfiles==0.21.0
weasyprint==61.0
whitenoise==6.6.0
